Added preview image conversion to update.py
A png or jpg in a post's images folder named "previewImage" gets
converted to "previewImageResized.png" by ImageMagick.
This allows the smaller image to be used when serving up a post preview,
saving bandwidth.

Also updated README.md to include dependencies, and added post 0001 to
ensure batch conversion is working.