Added plan of project in plan.md
I am probably going to do a heavy refactoring of the code in update.py,
if not completely delete all of it. The idea is that I will make a plan
of which code passes what to which other code, and then write unit tests
for this. This should mean LazyPyCMS is a model of good practises for
writing Python code.