Added initial HTML saving ability
HTML is now saved in the post's new folder, but the actual original contents of
the post are not saved. This needs to be added soon - probably in a
separate file called save_post.py.

Editing should also be added, so that
already written posts can be modified. This could fairly easily be
accomplished by adding some functionality to make_post.py